requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
trafilatura==1.12.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
tqdm==4.66.4
python-dotenv==1.0.1
pydantic==2.7.4
python-multipart==0.0.9
openai==1.37.1
orjson==3.10.3
